hpet.

+++ src/acpi.h
@@ -2,7 +2,7 @@
 
 #include <stdint.h>
 
-extern uint8_t* hpet_block;
+extern uint64_t* hpet_block;
 
 void read_rsdp();
 //uint8_t* get_hpet_block();
